Refactor ListenerRegistration creation
  * Pull activity-scoping out ListenerRegistrationImpl--it's a totally
    separate concern.
  * Make the relationship between the various levels of listener clearer
    in the places where we construct them.
